requirements.txt
fastapi 
uvicorn
langchain
langchain-community
sentence-transformers
faiss-cpu
langchain_ollama
PyYAML
pypdf
pydantic
scikit-learn
python-multipart
langchain_huggingface